clean code as widget
