+++ test/widget_test.dart
@@ -7,13 +7,13 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:json_convert_dart/json_to_dart_app.dart';
 
-import 'package:json_convert_dart/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(  JsonToDartApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
